Add start and end dates to snowflake-lite output
