+++ dumper/app/src/main/java/com/google/edwmigration/dumper/application/dumper/connector/snowflake/SnowflakePlanner.java
@@ -119,4 +119,36 @@ final class SnowflakePlanner {
       return HeaderTransformerUtil.toCamelCaseFrom(caseFormat);
     }
   }
+
+  Task<?> proceduresTask() {
+    String view = "SNOWFLAKE.ACCOUNT_USAGE.PROCEDURES";
+    String query =
+        String.format(
+            "SELECT procedure_language, procedure_owner, count(1) FROM %s GROUP BY ALL", view);
+    ImmutableList<String> header = ImmutableList.of("Language", "Owner", "Count");
+    return new LiteTimeSeriesTask("procedures.csv", query, header);
+  }
+
+  Task<?> timeRangeTask() {
+    String view = "SNOWFLAKE.ACCOUNT_USAGE.WAREHOUSE_EVENTS_HISTORY";
+    String query = String.format("SELECT min(timestamp), max(timestamp) FROM %s", view);
+    ImmutableList<String> header = ImmutableList.of("StartTime", "EndTime");
+    return new LiteTimeSeriesTask("time_range.csv", query, header);
+  }
+
+  Task<?> warehouseEventsTask() {
+    String query =
+        "SELECT event_name, cluster_number, warehouse_id, warehouse_name, count(1)"
+            + " FROM SNOWFLAKE.ACCOUNT_USAGE.WAREHOUSE_EVENTS_HISTORY GROUP BY ALL";
+    ImmutableList<String> header =
+        ImmutableList.of("Name", "Cluster", "WarehouseId", "WarehouseName", "Count");
+    return new LiteTimeSeriesTask("warehouse_events.csv", query, header);
+  }
+
+  Task<?> warehouseMeteringTask() {
+    String view = "SNOWFLAKE.ACCOUNT_USAGE.WAREHOUSE_METERING_HISTORY";
+    String query = String.format("SELECT warehouse_name, count(1) FROM %s GROUP BY ALL", view);
+    ImmutableList<String> header = ImmutableList.of("Name", "Count");
+    return new LiteTimeSeriesTask("warehouse_metering.csv", query, header);
+  }
 }
